+++ src/main/groovy/io/xh/hoist/admin/ClusterObjectsReport.groovy
@@ -1,21 +1,20 @@
 package io.xh.hoist.admin
 
-
 import io.xh.hoist.json.JSONFormat
 
 class ClusterObjectsReport implements JSONFormat {
 
     // List of all of the distributed object data from all of the instances in the cluster.
-    List<ClusterObjectInfo> info
+    List<Map> info
     Map<String, List<List<String>>> breaks
     Long startTimestamp
     Long endTimestamp
 
-    ClusterObjectsReport(Map args) {
-        info = args.info as List<ClusterObjectInfo>
-        breaks = createBreaks()
-        startTimestamp = args.startTimestamp as Long
-        endTimestamp = args.endTimestamp as Long
+    ClusterObjectsReport(List<Map> info, Long startTimestamp, Long endTimestamp) {
+        this.info = info
+        this.breaks = createBreaks()
+        this.startTimestamp = startTimestamp
+        this.endTimestamp = endTimestamp
     }
 
     Map formatForJSON() {
@@ -32,13 +31,19 @@ class ClusterObjectsReport implements JSONFormat {
     //------------------
     private Map<String, List<List<String>>> createBreaks() {
         Map<String, List<List<String>>> breaks = [:].withDefault { [] }
-        info.groupBy { it.name }.each { name, infoObjs ->
-            [infoObjs, infoObjs].eachCombination { a, b ->
-                if (a !== b && !a.isMatching(b)) {
-                    breaks[name].push([a.instanceName, b.instanceName])
+        info.groupBy { it.name }
+            .each { name, infoObjs ->
+                [infoObjs, infoObjs].eachCombination { a, b ->
+                    if (a !== b &&
+                        (
+                            a.comparableAdminStats != b.comparableAdminStats ||
+                            a.comparableAdminStats.any { a.adminStats[it] != b.adminStats[it] }
+                        )
+                    ) {
+                        breaks[name].push([a.instanceName, b.instanceName])
+                    }
                 }
             }
-        }
         return breaks
     }
 }