+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### ⚙️ Technical
 
 * Support for new consolidated clients tab in Hoist-react v73.
+* Harden `ClusterObjectReport` against issues with serialization.
 
 ## 29.2.0 - 2025-04-14